requirements.txt
# ------------------------------
# Core & Environment
# ------------------------------
pydantic>=2.0
python-dotenv>=1.0


# ------------------------------
# PDF Processing (Research Papers)
# ------------------------------
pdfplumber>=0.10.0
pypdf>=4.0


# ------------------------------
# NLP & Text Processing
# ------------------------------
nltk>=3.8
spacy>=3.7
scikit-learn>=1.3


# ------------------------------
# Embeddings & Vector Search
# ------------------------------
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4


# ------------------------------
# LLM / RAG (GROQ Compatible)
# ------------------------------
langchain>=0.1.0
langchain-community>=0.0.30
langchain-groq>=0.1.0


# ------------------------------
# External APIs / MCP Tools
# ------------------------------
requests>=2.31.0


# ------------------------------
# Data Analysis
# ------------------------------
pandas>=2.0
numpy>=1.24


# ------------------------------
# UI & Visualization
# ------------------------------
streamlit>=1.30
matplotlib>=3.7
plotly>=5.18
networkx>=3.2
